Rename: Collection of data-related files into the data folder
