+++ src/main/java/com/nimoh/hotel/commons/board/BoardErrorResult.java
@@ -1,4 +1,4 @@
-package com.nimoh.hotel.errors.board;
+package com.nimoh.hotel.commons.board;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
