+++ src/main/java/com/nimoh/hotel/controller/RoomController.java
@@ -1,7 +1,7 @@
 package com.nimoh.hotel.controller;
 
-import com.nimoh.hotel.dto.room.RoomDetailResponse;
-import com.nimoh.hotel.errors.room.RoomException;
+import com.nimoh.hotel.data.dto.room.RoomDetailResponse;
+import com.nimoh.hotel.commons.room.RoomException;
 import com.nimoh.hotel.service.room.RoomService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
