+++ src/main/java/com/nimoh/hotel/controller/ValidationController.java
@@ -1,14 +1,7 @@
 package com.nimoh.hotel.controller;
 
-import com.nimoh.hotel.dto.user.UserSignUpRequest;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-
-import javax.validation.Valid;
 
 @RestController
 @RequestMapping("/api/v1/validation")
