+++ src/main/java/com/nimoh/hotel/data/entity/Reservation.java
@@ -1,4 +1,4 @@
-package com.nimoh.hotel.domain;
+package com.nimoh.hotel.data.entity;
 
 import lombok.Data;
