+++ src/main/java/com/nimoh/hotel/repository/ReservationRepository.java
@@ -1,7 +1,7 @@
 package com.nimoh.hotel.repository;
 
 
-import com.nimoh.hotel.domain.Reservation;
+import com.nimoh.hotel.data.entity.Reservation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ReservationRepository extends JpaRepository<Reservation, Long> {
