+++ src/main/java/com/nimoh/hotel/repository/RoomRepository.java
@@ -1,12 +1,11 @@
 package com.nimoh.hotel.repository;
 
 
-import com.nimoh.hotel.domain.Room;
+import com.nimoh.hotel.data.entity.Room;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
-import java.util.Optional;
 
 @Repository
 public interface RoomRepository extends JpaRepository<Room, Long> {
