+++ src/main/java/com/nimoh/hotel/repository/UserRepository.java
@@ -1,8 +1,7 @@
 package com.nimoh.hotel.repository;
 
-import com.nimoh.hotel.domain.User;
+import com.nimoh.hotel.data.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
