+++ src/main/java/com/nimoh/hotel/service/room/RoomServiceImpl.java
@@ -1,9 +1,9 @@
 package com.nimoh.hotel.service.room;
 
-import com.nimoh.hotel.domain.Room;
-import com.nimoh.hotel.dto.room.RoomDetailResponse;
-import com.nimoh.hotel.errors.room.RoomErrorResult;
-import com.nimoh.hotel.errors.room.RoomException;
+import com.nimoh.hotel.data.entity.Room;
+import com.nimoh.hotel.data.dto.room.RoomDetailResponse;
+import com.nimoh.hotel.commons.room.RoomErrorResult;
+import com.nimoh.hotel.commons.room.RoomException;
 import com.nimoh.hotel.repository.RoomRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
