+++ src/main/java/com/nimoh/hotel/service/user/UserService.java
@@ -1,7 +1,7 @@
 package com.nimoh.hotel.service.user;
 
-import com.nimoh.hotel.dto.user.UserResponse;
-import com.nimoh.hotel.dto.user.UserSignUpRequest;
+import com.nimoh.hotel.data.dto.user.UserResponse;
+import com.nimoh.hotel.data.dto.user.UserSignUpRequest;
 
 public interface UserService {
     public UserResponse signUp(UserSignUpRequest request);
