+++ src/main/resources/mybatis/sql/Room.xml
@@ -1,13 +1,13 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
 <mapper namespace="com.nimoh.hotel.repository.RoomRepository">
-    <select id="getList" parameterType="com.nimoh.hotel.domain.Room" resultType="com.nimoh.hotel.domain.Room">
+    <select id="getList" parameterType="com.nimoh.hotel.data.entity.Room" resultType="com.nimoh.hotel.data.entity.Room">
         SELECT
         *
         FROM Room
     </select>
 
-    <select id="get" parameterType="int" resultType="com.nimoh.hotel.domain.Room">
+    <select id="get" parameterType="int" resultType="com.nimoh.hotel.data.entity.Room">
         SELECT
         *
         FROM Room
