+++ src/test/java/com/nimoh/hotel/repository/RoomRepositoryTest.java
@@ -1,6 +1,6 @@
 package com.nimoh.hotel.repository;
 
-import com.nimoh.hotel.domain.Room;
+import com.nimoh.hotel.data.entity.Room;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
