+++ src/test/java/com/nimoh/hotel/repository/UserRepositoryTest.java
@@ -1,7 +1,6 @@
 package com.nimoh.hotel.repository;
 
-import com.nimoh.hotel.domain.User;
-import org.assertj.core.api.Assertions;
+import com.nimoh.hotel.data.entity.User;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
